refactor: extract native playback suite helpers
